requirements.txt
asarPy==1.0.1
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
better-exceptions==0.3.3
boto3==1.35.57
botocore==1.35.57
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
colorama==0.4.6
confuse==2.0.1
constantly==23.10.4
cryptography==43.0.3
Cython==3.0.11
daphne==4.1.2
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
et_xmlfile==2.0.0
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
load-dotenv==0.1.0
loguru==0.7.2
lxml==5.3.0
Mako==1.3.8.dev0
Markdown==3.5.2
MarkupSafe==3.0.2
meson==1.5.2
numpy==2.1.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pg8000==1.31.2
pillow==11.0.0
pyactiveresource==2.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.27.0
pycparser==2.22
Pygments==2.18.0
PyGObject==3.50.0
PyJWT==2.9.0
pyOpenSSL==25.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-amazon-sp-api==1.6.85
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-aws4auth==1.3.1
s3transfer==0.10.3
scramp==1.4.5
service-identity==24.2.0
setuptools==75.6.0.post20241226
ShopifyAPI==12.6.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.2
tabulate==0.9.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.45.1
win32-setctime==1.1.0
zope.interface==7.2
